TSUNAMI-TREND: fractional shares, slice recalibrated 0.40 -> 0.30
Per Leron: whole shares on the $500 book could not express 6 of 7
trending signals (SPXL $275/share vs a ~$60 target weight). Fractional
shares fix that, but at SLICE=0.40 they over-deploy (~70% avg gross vs
~51%) and cost ~0.4 Sharpe in every backtest window. Slice sweep
{0.40,0.30,0.25,0.20,0.15} on walk-forward/recent/full windows: 0.30 is
the calibrated match -- Sharpe 1.23/1.47/1.27 vs whole-share's
1.43/1.67/1.44 with similar MaxDD, while expressing every signal. Index
sleeve keeps its 0.375 ratio (0.1125).

- shares columns INTEGER -> NUMERIC(16,6) (idempotent widening ALTERs in
  _DDL for existing installs; paper tables, trivially small)
- $1 minimum order notional (dust guard)
- routes/trend status: float share handling
- 18/18 tests (SLICE/override/param assertions updated)

Dry-run vs live book state: next rebalance deploys ~47% gross across 6
trending names (AMDL/SPXL/TQQQ/SBIT/ETHD/SMST) vs 12% today; NVDL
correctly excluded by the split-adjusted-history fix.

+++ spreadworks/backend/routes_tsunami.py
@@ -822,7 +822,7 @@ def tsunami_trend_status() -> dict[str, Any]:
 
     book_out, market_value, cost_basis = [], 0.0, 0.0
     for letf, shares, avg_cost, updated_at in book_rows:
-        shares = int(shares)
+        shares = float(shares)
         avg_cost = float(avg_cost)
         quote = tradier_client.get_quote(letf)
         last = float((quote or {}).get("last") or (quote or {}).get("close") or 0) or avg_cost
@@ -845,7 +845,7 @@ def tsunami_trend_status() -> dict[str, Any]:
         "realized_pnl": float(realized_total_rows[0][0]) if realized_total_rows else 0.0,
         "book": book_out,
         "recent_trades": [{"ts": str(r[0]), "letf": r[1], "side": r[2],
-                           "shares": int(r[3]), "price": float(r[4]), "reason": r[5],
+                           "shares": float(r[3]), "price": float(r[4]), "reason": r[5],
                            "realized_pnl": float(r[6]) if r[6] is not None else None}
                           for r in trades],
     }
